+++ changelog.rst
@@ -8,6 +8,7 @@ Last change: 05-JAN-2026 CEST
 Important updates:
 ^^^^^^^^^^^^^^^^^^
 - Localize GUI allows the user to select which localization columns to save when saving localizations. See the new dialog in the *File* -> *Select columns to save*
+- Localize accepts frame bounds to analyze only a subset of frames
 
 *Small improvements:*
 +++++++++++++++++++++